Rename "tool-inputs-from" to "input-ports-form". Rename "gui-editor" to "clt-editor". Move sidebar component outside the clt-editor dir.
