+++ client/src/app/components/clt-editor/animation.states.ts
