+++ client/src/app/components/clt-editor/clt-editor.component.scss
