+++ client/src/app/components/clt-editor/shared/gui-editor.events.ts
