+++ client/src/app/components/clt-editor/shared/sidebar.enums.ts
