+++ client/src/app/components/sidebar/editor-sidebar.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../assets/sass/variables";
+@import "../../../assets/sass/variables";
 
 .editor-sidebar-component {
     right: 0;
