+++ client/src/app/components/sidebar/object-inpsector/object-inpsector.component.scss
